+++ src/main/java/fi/jubic/quanta/models/Anomaly.java
@@ -5,14 +5,14 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import fi.jubic.easymapper.annotations.EasyId;
 import fi.jubic.easyvalue.EasyValue;
-import fi.jubic.quanta.db.tables.records.DetectionResultRecord;
+import fi.jubic.quanta.db.tables.records.AnomalyRecord;
+import fi.jubic.quanta.util.DateUtil;
 
-import java.sql.Timestamp;
 import java.time.Instant;
-import java.util.Arrays;
-import java.util.List;
+import java.util.Collections;
+import java.util.Map;
 
-import static fi.jubic.quanta.db.tables.DetectionResult.DETECTION_RESULT;
+import static fi.jubic.quanta.db.tables.Anomaly.ANOMALY;
 
 @EasyValue
 @JsonDeserialize(builder = Anomaly.Builder.class)
@@ -24,13 +24,13 @@ public abstract class Anomaly {
 
     public abstract Instant getEnd();
 
-    public abstract List<Measurement> getValues();
+    public abstract Map<String, Object> getSample();
 
     public abstract String getClassification();
 
     public abstract Double getProbability();
 
-    public abstract Delta getDeltaMax();
+    public abstract Map<String, Object> getMetadata();
 
     public abstract Builder toBuilder();
 
@@ -39,70 +39,55 @@ public abstract class Anomaly {
     }
 
     public static class Builder extends EasyValue_Anomaly.Builder {
-
+        @Override
+        public Anomaly.Builder defaults(Anomaly.Builder builder) {
+            return builder
+                    .setId(0L)
+                    .setMetadata(Collections.emptyMap());
+        }
     }
 
-    public static final AnomalyRecordMapper<DetectionResultRecord> mapper = AnomalyRecordMapper
-            .builder(DETECTION_RESULT)
-            .setIdAccessor(DETECTION_RESULT.ID)
+    public static final AnomalyRecordMapper<AnomalyRecord> mapper = AnomalyRecordMapper
+            .builder(ANOMALY)
+            .setIdAccessor(ANOMALY.ID)
             .setStartAccessor(
-                    DETECTION_RESULT.STARTING_TIME,
-                    Timestamp::from,
-                    Timestamp::toInstant
+                    ANOMALY.STARTING_TIME,
+                    DateUtil::toLocalDateTime,
+                    DateUtil::toInstant
             )
             .setEndAccessor(
-                    DETECTION_RESULT.ENDING_TIME,
-                    Timestamp::from,
-                    Timestamp::toInstant
+                    ANOMALY.ENDING_TIME,
+                    DateUtil::toLocalDateTime,
+                    DateUtil::toInstant
             )
-            .setValuesAccessor(
-                    DETECTION_RESULT.VALUES,
-                    Anomaly::valuesToString,
-                    Anomaly::valuesFromString
+            .setSampleAccessor(
+                    ANOMALY.SAMPLE,
+                    Anomaly::mapToString,
+                    Anomaly::mapFromString
             )
-            .setClassificationAccessor(DETECTION_RESULT.CLASSIFICATION)
-            .setProbabilityAccessor(DETECTION_RESULT.PROBABILITY)
-            .setDeltaMaxAccessor(
-                    DETECTION_RESULT.DELTA_MAX,
-                    Anomaly::deltaToString,
-                    Anomaly::deltaFromString
+            .setClassificationAccessor(ANOMALY.CLASSIFICATION)
+            .setProbabilityAccessor(ANOMALY.PROBABILITY)
+            .setMetadataAccessor(
+                    ANOMALY.METADATA,
+                    Anomaly::mapToString,
+                    Anomaly::mapFromString
             )
             .build();
 
     private static final ObjectMapper valueMapper = new ObjectMapper();
 
-    private static String valuesToString(List<Measurement> values) {
+    private static String mapToString(Map<String, Object> metadata) {
         try {
-            return valueMapper.writeValueAsString(values);
+            return valueMapper.writeValueAsString(metadata);
         }
         catch (JsonProcessingException e) {
             throw new IllegalStateException(e);
         }
     }
 
-    private static List<Measurement> valuesFromString(String string) {
+    private static Map<String, Object> mapFromString(String string) {
         try {
-            return Arrays.asList(
-                    valueMapper.readValue(string, Measurement[].class)
-            );
-        }
-        catch (JsonProcessingException e) {
-            throw new IllegalStateException(e);
-        }
-    }
-
-    private static String deltaToString(Delta delta) {
-        try {
-            return valueMapper.writeValueAsString(delta);
-        }
-        catch (JsonProcessingException e) {
-            throw new IllegalStateException(e);
-        }
-    }
-
-    private static Delta deltaFromString(String string) {
-        try {
-            return valueMapper.readValue(string, Delta.class);
+            return valueMapper.readValue(string, Map.class);
         }
         catch (JsonProcessingException e) {
             throw new IllegalStateException(e);