+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Rename `type` to `valueType` in column data models.
+- Update anomaly data model
 
 ## [0.3.5] - 2021-07-19
 ### Changed